Nonparametric Double Conditional Smoothing for 2D Surfaces

DCSmooth provides a double conditional nonparametric smoothing of the expectation surface of a functional time series or a random field on a lattice. Bandwidth selection is done via an iterative plug-in method.

dcs(Y, dcs_options = set.options(), X = 1, T = 1, h = "auto")

Arguments

Y
A numeric matrix that contains the observations of the random field or functional time-series.

dcs_options
An object of class "dcs_options", specifying the parameters for the smoothing and bandwidth selection procedure.

X
An optional numeric vector containing the exogenous covariates with respect to the rows.

T
An optional numeric vector containing the exogenous covariates with respect to the columns.

h
Bandwidth for smoothing the observations in Y. Can be a two-valued numerical vector with bandwidths in row- and column-direction. If the value is "auto" (the default), bandwidth selection will be carried out by the iterative plug-in algorithm.

Value

DCSmooth returns an object of class "dcs", including

Y	matrix of original observations.

X, T	vectors of covariates over rows (X) and columns (T).

M	resulting matrix of smoothed values.

R	matrix of resudials of estimation, \(Y - M\).

h	optimized or given bandwidths.

c_f	estimated variance coefficient.

dcs_options	an object of class cds_options containing the initial options of the dcs procedure.

iterations	number of iterations of the IPI-procedure.

time_used	time spend searching for optimal bandwidths (not overall runtime of the function).

qarma	optional return, if method "qarma" is chosen for estimation of the variance factor. Omitted, if "iid" is used.

dcs_options	an object of class cds_options containing the initial options of the dcs procedure.

Details

See the vignette for a more detailed description of the function.

See also

Examples

y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs(y)

#> dcs 
#> -------------------------------------- 
#> DCS with automatic bandwidth selection
#> -------------------------------------- 
#>  Selected Bandwidths:
#> 		h_x: 0.20219 
#> 		h_t: 0.14973 
#>  Variance Factor:
#> 		c_f: 0.974 
#> --------------------------------------