Print Results from Double Conditional Smoothing

print method for class "dcs"

# S3 method for dcs
print(object)

Arguments

object
an object of class "dcs", usually, a result of a call to dcs.

Details

print.dcs prints a short summary of an object of class dcs, only including bandwidths and the estimated variance coefficient (only if automatic bandwidth selection is used).

See also

Examples

y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs_object <- dcs(y)
print(dcs_object)

#> dcs 
#> -------------------------------------- 
#> DCS with automatic bandwidth selection
#> -------------------------------------- 
#>  Selected Bandwidths:
#> 		h_x: 0.18339 
#> 		h_t: 0.18683 
#>  Variance Factor:
#> 		c_f: 0.99117 
#> -------------------------------------- 

dcs_object

#> dcs 
#> -------------------------------------- 
#> DCS with automatic bandwidth selection
#> -------------------------------------- 
#>  Selected Bandwidths:
#> 		h_x: 0.18339 
#> 		h_t: 0.18683 
#>  Variance Factor:
#> 		c_f: 0.99117 
#> --------------------------------------