Estimation of a QARMA-process

Parametric Estimation of a \(QARMA(p, q)\)-process on a lattice using the Hannan-Rissanen algorithm.

qarma.est(Y, model_order = list(ar = c(1, 1), ma = c(1, 1)))

Arguments

Y
A numeric matrix that contains the demeaned observations of the random field or functional time-series.

model_order
A list containing the orders of the QARMA model in the form model_order = list(ar = c(p1, p2), ma = c(q1, q2)). Default value is a \(QARMA((1, 1), (1, 1))\) model.

Value

The function returns a list including

ar	a \((p_1 + 1) \times (p_2 + 1)\)-matrix containing the estimated AR-parameters \(\phi\) of the QARMA-process.

ma	a \((q_1 + 1) \times (q_2 + 1)\)-matrix containing the estimated MA-parameters \(\theta\) of the QARMA-process.

sigma	the estimated standard deviation \(\sigma\) of the QARMA-model.

innov	the matrix of innovations resulting from the QARMA estimation.

Details

The MA- and AR-parameters of a top-left quadrant ARMA process are estimated using the Hannen-Rissanen Algorithm (Hannen-Rissanen 1992). The lag-orders of the \(QARMA(p, q)\) are given by \(p = (p_1, p_2), q = (q_1, q_2)\), where \(p_1, q_1\) are the lags over the rows and \(p_2, q_2\) are the lags over the columns. The estimation process is based on the model $$\phi(B_{1}B_{2})X_{i,j} = \theta(B_{1}B_{2})u_{i,j}$$

See also

Examples