Simulation of a \(QARMA(p, q)\)-process

qarma.sim is used to simulate a specified QARMA-model on a lattice, with normally distributed innovations.

qarma.sim(n_x, n_t, model = list(ar, ma, sigma))

Arguments

Y
A numeric matrix that contains the demeaned observations of the random field or functional time-series.

model_order
A list containing the orders of the QARMA model in the form model_order = list(ar = c(p1, p2), ma = c(q1, q2)). Default value is a \(QARMA((1, 1), (1, 1))\) model.

Value

The function returns a list, consisting of

ar	a \((p_1 + 1) \times (p_2 + 1)\)-matrix containing the given AR-parameters \(\phi\) of the QARMA-process.

ma	a \((q_1 + 1) \times (q_2 + 1)\)-matrix containing the given MA-parameters \(\theta\) of the QARMA-process.

sigma	the standard deviation \(\sigma\) of the innovations.

innov	the simulated matrix of innovations.

Details

The values of a \(QARMA((p_2, p_2), (q_1, q_2))\)-model are simulated from given matrices \(\phi\) (ar) and \(\theta\) (ma) and a standard deviation for the innovations \(\sigma\) (sigma). The simulation process is based on the model $$\phi(B_{1}B_{2})X_{i,j} = \theta(B_{1}B_{2})u_{i,j}$$ where the innovations \(u\) are iid. draws from a standard normal distribution with zero mean and variance \(\sigma^2\)

see the vignette for further details.

See also

Examples

ma = matrix(c(1, 0.2, 0.4, 0.1), nrow = 2, ncol = 2)
ar = matrix(c(1, 0.5, -0.1, 0.1), nrow = 2, ncol = 2)
sigma = 0.5
q_model = list(ar = ar, ma = ma, sigma = sigma)

q_sim = qarma.sim(100, 100, model = model)

#> Error in as.matrix(model$ar): Objekt 'model' nicht gefunden

q_sim

#> Error in eval(expr, envir, enclos): Objekt 'q_sim' nicht gefunden