Set Options for the DCS procedure

either local polynomial regression ("LP", the default) or kernel regression ("KR").

set.options(...)

Arguments

kern_par
integer vector of length 2, containing the kernel parameters \(k\) and \(\mu\). Default is c(2, 0).

drv
A non-negative vector of length 2, containing the derivative orders to be estimated from the given data. Default ist c(0, 0). For LP-regression, polynomial order is selected as \((\nu_1 + 1, \nu_2 + 1)\).

infl_exp
A vector for the inflation exponents for bandwidth selection, with default values depending on the regression type.

infl_par
A vector for the inflation parameters for bandwidth selection, with default values depending on the regression type.

delta
A vector indicating the percentage of cutting off boundaries. Default for kernel regression is 5 polynomial regression.

const_window
logical. Should a constant window be used at the boundaries? Default is FALSE.

var_est
the method of estimating the variance coefficient \(c_f\). Currently available are var_est = "iid" (the default), where an iid. error strucutre is modeled and var_est = "qarma", where a spatial ARMA model is applied.

qarma_order
A list containing the order of the qarma model to be estimated. Default is a \(QARMA((1, 1), (1, 1))\)-model (qarma_model = list(ar = c(1, 1), ma = c(1, 1))). Could be also one of c("gpac", "bic") for automatic order selection. If var_est = "iid", this argument is ignored.

order_max
A list specifying the maximum order, when automatic order selection is used. Default is order_max = list(ar = c(1, 1), ma = c(1, 1)).

Value

An object of class "dcs_options".

Details

This function is used to set the options for bandwidth selection in the dcs function. Detailed information can be found in the vignette

See also

Examples

set.options()

#> dcs_options 
#> --------------------------------------- 
#> options for DCS 	 rows 	 cols 
#> --------------------------------------- 
#> type: local polynomial regression 
#> kernel order: 	 	  2 	 2 
#> derivative: 	 	  0 	 0 
#> polynomial order: 	  1 	 1 
#> inflation exponent: 	  0.6 	 0.6 
#> inflation parameters: 	  1.5 	 1 
#> boundary factor: 	  0 	 0 
#> constant window: 	  FALSE 
#> --------------------------------------- 
#> variance estimation: 	  iid. 
#> --------------------------------------- 

myOpt <- set.options(type = "KR", var_est = "qarma")
y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs(y, dcs_options = myOpt)

#> Warning: QARMA model not stationary, try another order for the AR-parts.

#> Warning: QARMA model not stationary, try another order for the AR-parts.

#> Warning: QARMA model not stationary, try another order for the AR-parts.

#> Warning: QARMA model not stationary, try another order for the AR-parts.

#> dcs 
#> -------------------------------------- 
#> DCS with automatic bandwidth selection
#> -------------------------------------- 
#>  Selected Bandwidths:
#> 		h_x: 0.16149 
#> 		h_t: 0.19013 
#>  Variance Factor:
#> 		c_f: 1.04276 
#> --------------------------------------