Summarizing Results from Double Conditional Smoothing

summary method for class "dcs"

# S3 method for dcs
summary(object)

# S3 method for summary_dcs
print(object, digits = max(3, getOption("digits") - 3))

Arguments

object
an object of class "dcs", usually, a result of a call to dcs.

digits
the number of significant digits to use when printing.

Value

The function summary.dcs returns an object of class summary_dcs including

h_opt	estimated optimal bandwidth from the IPI-procedure.

c_f	estimated variance factor.

iterations	number of iterations of the IPI-procedure.

time_used	time spend searching for optimal bandwidths (not overall runtime of the function).

qarma	optional return, if method "qarma" is chosen for estimation of the variance factor. Omitted, if "iid" is used.

dcs_options	an object of class cds_options containing the initial options of the dcs procedure.

Details

summary.dcs strips an object of class "dcs" from all large matrices (Y, X, T, M, R), allowing for easier handling of meta-statistics of the bandwidth selection procedure.

print.summary_dcs returns a list of summary statistics from the dcs procedure. The output depends on the use of the dcs- function. If automatic bandwidth selection is chosen, summary.dcs prints detailed statistics of the type of regression, the estimated bandwidths h_x, h_t, the variance coefficient c_f and performance statistics such as the number of iterations of the IPI-algorithm and the time used for bandwidth selection.

The method used for estimation of the variance coefficient is printed and the results of an QARMA-estimation, if available.

If bandwidths are supplied to dcs, summary.dcs only prints the given bandwidths.

Examples

y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs_object <- dcs(y)
summary(dcs_object)

#> summary_dcs 
#> ------------------------------------------ 
#> DCS with automatic bandwidth selection:
#> ------------------------------------------ 
#> Results of local polynomial regression:
#> Estimated Bandwidths: 	 h_x:	 0.1805 
#> 	 	 	 h_t:	 0.2006 
#> Variance Factor: 	 c_f:	 0.9913 
#> Iterations:			 9 
#> Time used (seconds):	 	 0.096 
#> ------------------------------------------ 
#> Variance Estimation: 	 iid.
#> ------------------------------------------ 
#> See used parameter with "$dcs_options".
#> ------------------------------------------